Fixed a bug that caused timing problems after x time

+++ Software/Firmware/smart-sensor/include/boards/Board.h
@@ -65,7 +65,7 @@ protected:
     SmartSensorBoard(): totalDrivers(0) {}
 
     /* Hold the timestamp in seconds where we have got a measurement. */
-    uint16_t measurementReceivedTimestamp;
+    uint32_t measurementReceivedTimestamp;
 
     /* Hold the amount of measurements that have been send out! */
     uint8_t totalMeasurementsSend;
